Create Bank_Credit_fraud_Analysis(EDA)
Credit-lending project is done to analyze the patterns present in the Bank loan data so that bank can analyze and indicate if a client has difficulty paying instalments, which may be used for taking actions such as denying the loan, reducing the amount of loan, lending (to risky applicants) at a higher interest rate, etc. Bank want to reduce their default exposure in terms of credit lending to customers.